feat: Split into packages
