+++ src/main/java/bob/exceptions/IllegalCommandException.java
@@ -1,3 +1,5 @@
+package bob.exceptions;
+
 public class IllegalCommandException extends Exception {
     public IllegalCommandException(String message) {
         super("-----------------------------\n" + message + "\n-----------------------------");
