+++ src/main/java/bob/storage/Storage.java
@@ -1,3 +1,7 @@
+package bob.storage;
+
+import bob.task.*;
+
 import java.nio.file.Path;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
