+++ src/main/java/bob/task/Todo.java
@@ -1,0 +1,17 @@
+package bob.task;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    @Override
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+    @Override
+    public String toFileString() {
+        return "T | " + (isComplete ? "Y" : "N") + " | " + description;
+    }
+}
