+++ text-ui-test/runtest.bat
@@ -14,8 +14,8 @@ IF ERRORLEVEL 1 (
 )
 REM no error here, errorlevel == 0
 
-REM run the program, feed commands from input.txt file and redirect the output to the ACTUAL.TXT
-java -classpath ..\bin Bob < input.txt > ACTUAL.TXT
+REM run the program, feed bob.command from input.txt file and redirect the output to the ACTUAL.TXT
+java -classpath ..\bin bob.ui.Bob < input.txt > ACTUAL.TXT
 
 REM compare the output to the expected output
 FC ACTUAL.TXT EXPECTED.TXT
